Go: Move MigrateKeys to standalone client only; remove CHANGELOG entry
