+++ go/interfaces/generic_base_commands.go
@@ -127,33 +127,6 @@ type GenericBaseCommands interface {
 		migrateOptions options.MigrateOptions,
 	) (string, error)
 
-	// MigrateKeys atomically transfers the specified keys from a source Valkey instance to a
-	// destination Valkey instance.
-	//
-	// [valkey.io]: https://valkey.io/commands/migrate/
-	MigrateKeys(
-		ctx context.Context,
-		host string,
-		port int64,
-		keys []string,
-		destinationDB int64,
-		timeout int64,
-	) (string, error)
-
-	// MigrateKeysWithOptions atomically transfers the specified keys from a source Valkey instance to a
-	// destination Valkey instance with additional options.
-	//
-	// [valkey.io]: https://valkey.io/commands/migrate/
-	MigrateKeysWithOptions(
-		ctx context.Context,
-		host string,
-		port int64,
-		keys []string,
-		destinationDB int64,
-		timeout int64,
-		migrateOptions options.MigrateOptions,
-	) (string, error)
-
 	UpdateConnectionPassword(ctx context.Context, password string, immediateAuth bool) (string, error)
 
 	ResetConnectionPassword(ctx context.Context) (string, error)
